--- src/main/java/commands/AddCommand.java
@@ -1,5 +1,0 @@
-package commands;
-
-public abstract class AddCommand implements Command {
-
-}
