--- src/main/java/commands/Command.java
@@ -1,7 +1,0 @@
-package commands;
-
-
-public interface Command {
-
-    void execute();
-}
