--- src/main/java/commands/DeleteCommand.java
@@ -1,5 +1,0 @@
-package commands;
-
-public abstract class DeleteCommand implements Command {
-
-}
