--- src/main/java/commands/EditCommand.java
@@ -1,5 +1,0 @@
-package commands;
-
-public abstract class EditCommand implements Command {
-
-}
