--- src/main/java/commands/ExitCommand.java
@@ -1,7 +1,0 @@
-package commands;
-
-public class ExitCommand implements Command {
-    public void execute() {
-        System.exit(0);
-    }
-}
