--- src/main/java/exceptions/parser/InputFormatException.java
@@ -1,7 +1,0 @@
-package exceptions.parser;
-
-public class InputFormatException extends ParserException {
-    public InputFormatException(String message) {
-        super(message);
-    }
-}
