--- src/main/java/exceptions/parser/InputNotRecognisedException.java
@@ -1,7 +1,0 @@
-package exceptions.parser;
-
-public class InputNotRecognisedException extends ParserException {
-    public InputNotRecognisedException(String message) {
-        super(message);
-    }
-}
