--- src/main/java/exceptions/storage/LoadingException.java
@@ -1,7 +1,0 @@
-package exceptions.storage;
-
-public class LoadingException extends StorageException {
-    public LoadingException(String message) {
-        super(message);
-    }
-}
