--- src/main/java/exceptions/storage/SavingException.java
@@ -1,7 +1,0 @@
-package exceptions.storage;
-
-public class SavingException extends StorageException {
-    public SavingException(String message) {
-        super(message);
-    }
-}
