+++ src/main/java/numerator/Storage.java
@@ -1,5 +1,7 @@
-import exceptions.storage.LoadingException;
-import exceptions.storage.SavingException;
+package numerator;
+
+import numerator.exceptions.storage.LoadingException;
+import numerator.exceptions.storage.SavingException;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
