+++ src/main/java/numerator/Ui.java
@@ -1,5 +1,7 @@
-import exceptions.NumeratorException;
-import exceptions.storage.LoadingException;
+package numerator;
+
+import numerator.exceptions.NumeratorException;
+import numerator.exceptions.storage.LoadingException;
 
 public class Ui {
