+++ src/main/java/numerator/exceptions/parser/InputFormatException.java
@@ -1,0 +1,7 @@
+package numerator.exceptions.parser;
+
+public class InputFormatException extends numerator.exceptions.parser.ParserException {
+    public InputFormatException(String message) {
+        super(message);
+    }
+}
