+++ src/main/java/numerator/exceptions/parser/InputNotRecognisedException.java
@@ -1,0 +1,7 @@
+package numerator.exceptions.parser;
+
+public class InputNotRecognisedException extends numerator.exceptions.parser.ParserException {
+    public InputNotRecognisedException(String message) {
+        super(message);
+    }
+}
