+++ src/main/java/numerator/exceptions/storage/LoadingException.java
@@ -1,0 +1,7 @@
+package numerator.exceptions.storage;
+
+public class LoadingException extends numerator.exceptions.storage.StorageException {
+    public LoadingException(String message) {
+        super(message);
+    }
+}
