+++ src/main/java/numerator/exceptions/storage/SavingException.java
@@ -1,0 +1,7 @@
+package numerator.exceptions.storage;
+
+public class SavingException extends numerator.exceptions.storage.StorageException {
+    public SavingException(String message) {
+        super(message);
+    }
+}
