+++ src/main/java/numerator/task/Task.java
@@ -1,4 +1,4 @@
-package task;
+package numerator.task;
 
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
