+++ src/main/java/numerator/task/ToDos.java
@@ -1,4 +1,4 @@
-package task;
+package numerator.task;
 
 public class ToDos extends Task {
     public ToDos(String description) {
